automation/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
selenium>=4.15.0
playwright>=1.40.0
python-docx>=0.8.11
PyPDF2>=3.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-dateutil>=2.8.2
openpyxl>=3.1.0
aiohttp>=3.9.0
fake-useragent>=1.4.0
undetected-chromedriver>=3.5.0
webdriver-manager>=4.0.0
asyncio-throttle>=1.0.2
nest-asyncio>=1.5.0
